HBSD: Support toggling exploit mitigations on a per-binary basis
This commit adds the following variables that ports' entries can use:

1. PAGEEXEC_DISABLE
2. MPROTECT_DISABLE
3. SEGVGUARD_DISABLE
4. ASLR_DISABLE
5. DISALLOW_MAP32BIT_DISABLE
6. SHLIBRANDOM_DISABLE

Certain ports violate various exploit mitigations, most notably PaX
NOEXEC. With this commit, packages can now be shipped with those exploit
mitigations disabled such that users do not have to perform the toggle
themselves.

To use this feature, in the port's Makefile, set the desired variable.
For example, in an upcoming commit to www/firefox/Makefile, the
following will be set:

MPROTECT_DISABLE=lib/firefox/firefox lib/firefox/plugin-container

Note that using this functionality requires building the port as root,
since this form of toggling uses the privileged system filesystem
extended attribute namespace (see extattr(9)).

As of this commit, if tmpfs is used, this new feature will only work on
13-CURRENT as the updates to tmpfs has not been MFC'd to 12-STABLE. I am
waiting on a full package build with this feature to be performed for
stress testing. Otherwise, if the package build is done on UFS or ZFS,
this feature will be in effect.

Ideally, we would have some way to tell the user via `pkg info` or `pkg
query` which exploit mitigations have been toggled. More research and
development needs to happen in that area.

Note that pkg itself does not support setting the extended attributes,
so this work only works with non-Poudriere setups (ie, only `make
install` is supported at the moment.) It is likely that pkg itself needs
to be updated to instruct libarchive to set the filesystem extended
attributes.

Signed-off-by:	Shawn Webb <shawn.webb@hardenedbsd.org>